Updated gradient heading to be presentational feature
